+++ resources/views/layouts/login.blade.php
@@ -7,12 +7,12 @@
 			<div class="row justify-content-center">
 				<div class="col-md-8 col-lg-6 mt-md-4">
 
-					<img src="{{ asset('/img/logo_login.png') }}" class="img-fluid text-center my-sm-0 p-4" />
+					<div class="px-5 px-sm-0"><img src="{{ asset('/img/logo_login.png') }}" class="img-fluid text-center my-sm-0 p-4" /></div>
 
 					<div class="card mb-4">
 						<div class="card-body p-md-5">
 
-							<h1 class="display-4 text-center mb-3 mb-md-5">@yield('title')</h1>
+							<h1 class="display-4 text-center mb-4 mb-md-5">@yield('title')</h1>
 
 							@yield('content')