+++ app/Http/ViewComposers/ContextMenuComposer.php
@@ -580,6 +580,24 @@ class ContextMenuComposer {
                         'authorized' => Auth::user()->can('view', $donor)
                     ]
                 ];
+            case 'donations.edit':
+                $donor = $view->getData()['donor'];
+                $donation = $view->getData()['donation'];
+                return [
+                    'delete' => [
+                        'url' => route('donations.destroy', [$donor, $donation]),
+                        'caption' => __('app.delete'),
+                        'icon' => 'trash',
+                        'authorized' => Auth::user()->can('delete', $donation),
+                        'confirmation' => __('donations.confirm_delete_donation')
+                    ],
+                    'back' => [
+                        'url' => route('donors.show', $donor),
+                        'caption' => __('app.close'),
+                        'icon' => 'times-circle',
+                        'authorized' => Auth::user()->can('view', $donor)
+                    ]
+                ];
 
             //
             // Reporting