+++ Changelog.md
@@ -3,6 +3,8 @@
 ## NEXT
 
 * Updated Laravel to v5.6.9.
+* Added checkbox for remembering the session to login view, default is false.
+* Optimized login view for small screens.
 
 ## 1.11.2